% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_overlapping_area.r
\name{fix_overlapping_area}
\alias{fix_overlapping_area}
\alias{fix_overlapping_area}
\alias{fix_overlapping_area.sfc}
\alias{fix_overlapping_area}
\alias{fix_overlapping_area.sf}
\title{Fix overlapping areas error.}
\usage{
fix_overlapping_area(x, ...)

\method{fix_overlapping_area}{sfc}(x, ...)

\method{fix_overlapping_area}{sf}(x, ...)
}
\arguments{
\item{x}{simple features object}

\item{...}{arguments passed to methods}
}
\value{
simple features object
}
\description{
The object is decomposed to its segments, and triangulated. This
inserts new vertices where ever the segments overlap (invalidly)
and returns component triangles. The final fix is to union those triangles, which automatically
rebuilds the right paths.
}
